include also https://github.com/eclipse/xtext/pull/3206
